manual_version/pyproject.toml
[project]
name = "aitu-isscv-assignment3"
version = "0.1.0"
description = "Вариант практической работы шестой (6) - Телекоммуникации регистрация клиентов и выставление счетов"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "alembic>=1.13.0",
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
    "ruff>=0.15.10",
    "mypy>=1.20.1",
    "cryptography>=46.0.7",
]

[tool.mypy]
mypy_path = "typings"
